Refactor project fixtures
